# api-docs/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "GSES2 BTC application"
host: "localhost:3000"
basePath: "/api"
tags:
- name: "rate"
  description: "Отримання поточного курсу BTC до UAH"
- name: "subscription"
  description: "Робота з підпискою"
schemes:
- "http"
paths:
  /rate:
    get:
      tags:
      - "rate"
      summary: "Отримати поточний курс BTC до UAH"
      description: "Запит має повертати поточний курс BTC до UAH використовуючи будь-який third party сервіс з публічним АРІ"
      operationId: "rate"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Повертається актуальний курс BTC до UAH"
          schema:
            type: "number"
        "400":
          description: "Invalid status value"
        "500":
          description: "Internal server error"
  /subscribe:
    post:
      tags:
      - "subscription"
      summary: "Підписати емейл на отримання поточного курсу"
      description: "Запит має перевірити, чи немає данної електронної адреси в поточній базі даних (файловій) і, в разі її відсутності, записувати її. Пізніше, за допомогою іншого запиту ми будемо відправляти лист на ті електронні адреси, які будуть в цій базі. "
      operationId: "subscribe"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "email"
        in: "formData"
        description: "Електронна адреса, яку потрібно підписати"
        required: true
        type: "string"
      responses:
        "200":
          description: "E-mail додано"
        "409":
          description: "Повертати, якщо e-mail вже є в базі даних (файловій)"
        "500":
          description: "Internal server error"
  /sendEmails:
    post:
      tags:
      - "subscription"
      summary: "Відправити e-mail з поточним курсом на всі підписані електронні пошти."
      description: "Запит має отримувати актуальний курс BTC до UAH за допомогою third-party сервісу та відправляти його на всі електронні адреси, які були підписані раніше.  "
      operationId: "sendEmails"
      produces:
      - "application/json"
      responses:
        "200":
          description: "E-mailʼи відправлено"
        "400":
          description: "Error"
        "500":
            description: "Internal server error"